api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
transformers==4.35.0
ultralytics==8.0.186
opencv-python==4.8.1
python-multipart==0.0.6
scikit-learn==1.3.2